+++ lib/jekyll/tags/highlight.rb
@@ -41,8 +41,11 @@ eos
       end
 
       def render(context)
-        if context.registers[:site].pygments
+        case context.registers[:site].highlighter
+        when 'pygments'
           render_pygments(context, super)
+        when 'rouge'
+          render_rouge(context, super)
         else
           render_codehighlighter(context, super)
         end
@@ -58,9 +61,28 @@ eos
           @lang
         )
 
-        output = context["pygments_prefix"] + output if context["pygments_prefix"]
-        output = output + context["pygments_suffix"] if context["pygments_suffix"]
-        output
+        output = context["highlighter_prefix"] + output if context["highlighter_prefix"]
+        output << context["highlighter_suffix"] if context["highlighter_suffix"]
+
+        return output
+      end
+
+      def render_rouge(context, code)
+        require 'rouge'
+
+        linenos = @options.keys.include?('linenos')
+        lexer = Rouge::Lexer.find_fancy(@lang, code) || Rouge::Lexers::PlainText
+        formatter = Rouge::Formatters::HTML.new(line_numbers: linenos, wrap: false)
+
+        pre = "<pre>#{formatter.format(lexer.lex(code))}</pre>"
+
+        output = context["highlighter_prefix"] || ""
+        output << "<div class=\"highlight\">"
+        output << add_code_tags(pre, @lang)
+        output << "</div>"
+        output << context["highlighter_suffix"] if context["highlighter_suffix"]
+
+        return output
       end
 
       def render_codehighlighter(context, code)